+++ src/BDSSDType.cc
@@ -37,6 +37,8 @@ std::map<BDSSDType, std::string>* BDSSDType::dictionary =
 					 {BDSSDType::worldcomplete,          "worldcomplete"},
 					 {BDSSDType::collimator,             "collimator"},
 					 {BDSSDType::collimatorcomplete,     "collimatorcomplete"},
-		             {BDSSDType::apertureimpacts,        "aperture"},
-					 {BDSSDType::aperturecomplete,       "aperturecomplete"}
+					 {BDSSDType::apertureimpacts,        "aperture"},
+					 {BDSSDType::aperturecomplete,       "aperturecomplete"},
+					 {BDSSDType::thinthing,              "thinthing"},
+					 {BDSSDType::wirecomplete,           "wirecomplete"}
     });